Extract `ExitPoint` from `ControlFlowGraph`
`ExitPoint` class is not related to `ControlFlowGraph` class,
so they should be separated to avoid confusing
